FAQ refactoring rebase
